Ensure screenshot images are scalable
This fixes a bug I found because of the width of the
Flight Booker image. As I reduced the width of the browser
the right column scaled down but the left column remained
the same width. Thus the two columns didn't have the same
width anymore. This happened because the Flight Booker image
was not scalable. Setting the width to be 100% fixed it.